Updated Event Manager.
