+++ EventManager/IEvent.cs
@@ -1,0 +1,11 @@
+namespace Laserbean.EventManager
+{
+    /// <summary>
+    /// Interface for event classes. Unsure if this is necessary, but it is used to enforce a common structure for all events.
+    /// </summary>
+    public interface IEvent
+    {
+  
+    }
+
+}
